# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Saray Property Booking System",
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "URL"
        }
    ],
    "paths": {
...
        "/api/buildings": {
            "get": {
                "tags": [
                    "Building"
                ],
                "summary": "List all buildings with optional filters and pagination",
                "description": "Display a paginated listing of buildings.",
                "operationId": "704d1f7897ed7a3b8d29dacea66823ad",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter by name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Filter by location",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Items per page (max 100)",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of buildings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Building"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden — insufficient permissions"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Building"
                ],
                "summary": "Create a new building",
                "description": "Store a newly created building in storage, optionally with an image.",
                "operationId": "c8f8399919e01b1e4e2e2f6a9f1f5107",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location",
                                    "status",
                                    "ecd"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Building A"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Downtown"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Off-Plan"
                                    },
                                    "ecd": {
                                        "type": "string",
                                        "example": "Q4-2026"
                                    },
                                    "image": {
                                        "description": "Optional building image (jpeg, png, gif)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Building created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Building"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Forbidden — insufficient permissions"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/buildings/{buildingId}/units": {
            "get": {
                "tags": [
                    "Building"
                ],
                "summary": "List units in a building",
                "description": "List all units for a specific building with optional filters.",
                "operationId": "2ea5af369ed04223da6d7b2fb00aff75",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "description": "Building ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "unit_no",
                        "in": "query",
                        "description": "Filter by unit number",
                        "schema": {
                            "type": "string",
                            "example": "A101"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Pending",
                                "Available",
                                "Pre-Booked",
                                "Booked",
                                "Sold",
                                "Pre-Hold",
                                "Hold",
                                "Cancelled"
                            ],
                            "example": "Available"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Items per page (max 100)",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of units",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Unit"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden — insufficient permissions"
                    },
                    "404": {
                        "description": "Not Found — building does not exist"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Building": {
                "title": "Building",
                "required": [
                    "id",
                    "name",
                    "location",
                    "status",
                    "ecd",
                    "added_by_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Building A"
                    },
                    "location": {
                        "type": "string",
                        "example": "Downtown"
                    },
                    "status": {
                        "type": "string",
                        "example": "Off-Plan"
                    },
                    "ecd": {
                        "type": "string",
                        "example": "Q4-2026"
                    },
                    "added_by_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-02T00:00:00Z"
                    },
                    "image_url": {
                        "description": "URL to fetch the building’s image",
                        "type": "string",
                        "format": "url",
                        "example": "https://your-domain.com/api/buildings/1/image"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}